client/render: the zoom is a ratio, and the inverse the camera never had
The camera could turn a tile into pixels and had no way back, so nothing
could answer which tile is under the cursor -- which a zoom about the
cursor, a drag and eventually picking all reduce to. `ScreenPoint` was
also two spaces under one name: world pixels out of `project`, image
pixels out of `to_screen`, and no type able to tell them apart.

So `WorldPixel` and `ViewPixel`, neither with `From` or `Into` -- the only
thing allowed to move between them is a camera, and a conversion that
needs a camera is a method. `unproject` is `project`'s named inverse,
exact, and it takes the height rather than guessing one: a pixel is a
column of tiles and that is the whole difficulty of picking. The eye is a
private `WorldPixel` behind `look_at`, because two writers already fought
over where the camera looks.

`Zoom` is a fraction from a fixed ladder and not an `f32`: `Camera` stays
`Copy + Eq` for the tests that compare cameras, and the offscreen target's
size comes out the same integer every frame instead of wobbling on
rounding noise. The scale is applied in exactly one place -- the world is
drawn at 1:1 into a texture of `ceil(viewport / zoom)` and `blit.rs` puts
that on the surface -- so no world pass learns what a zoom is and every
pixel-exact assertion downstream keeps meaning what it meant. Nearest
above 1, linear below: magnified, a texel has to stay a square; minified,
nearest samples one texel in four and the ground shimmers.

In the app: the wheel zooms about the cursor, a middle-drag pans, `Home`
re-locks to the body. The drag's fraction of a world pixel is accumulated
in the input handler and never in the eye -- an eye carrying it would put
every sprite on a half-texel boundary for half of all camera positions.
Zooming out is refused, and reported, when the offscreen texture would
exceed `max_texture_dimension_2d`: a truncated target draws a smaller
world into a larger rect, which reads as a projection bug.

The bounds tests are re-run at every rung rather than asserting the
constant that happened to hold at 1:1, and the frame test asserts the blit
at zoom 1 is the world image texel for texel -- without which every other
pixel assertion here measures an image the screen never shows.

+++ crates/client/render/src/depth.rs
@@ -81,8 +81,13 @@ impl Order {
 }
 
 /// The tile depth a camera centre sits at — the `base` for [`Order::to_depth`].
-pub fn base_for(x: u16, y: u16) -> i32 {
-    i32::from(x) + i32::from(y)
+///
+/// `i32` and not `u16`: a free camera is a pixel and not a tile, so the tile it
+/// is over comes from [`crate::camera::unproject`] and may be off the map
+/// entirely. That is harmless here — the base only says where the middle of the
+/// depth range is, and `DEPTH_TILES` is 512 either side of it.
+pub fn base_for(x: i32, y: i32) -> i32 {
+    x + y
 }
 
 /// Where a land tile sorts.
